Separate out types package into its own separate package from prisma-client
This allows the prisma-client package to stay purely focused on only exporting generated Prisma code for use elsewhere, while the types package can be its own thing. This seems like a more maintainable solution, especially as more resources get added and I'd need the better separation of concerns. This way, the prisma-client package contains minimal TypeScript while the types package has a lot more.
