+++ apps/back-end/package.json
@@ -23,6 +23,7 @@
   "type": "module",
   "dependencies": {
     "@neurosongs/prisma-client": "^1.0.0",
+    "@neurosongs/types": "^1.0.0",
     "@prisma/client": "^6.15.0",
     "dotenv": "^17.2.1",
     "express": "^5.1.0",
