--- packages/prisma-client/eslint.config.js
@@ -1,11 +1,0 @@
-import plugin from "@alextheman/eslint-plugin";
-
-export default [
-  ...plugin.configs.alexTypeScriptBase,
-  {
-    rules: {
-      // Imports in this package need to be relative so that they resolve correctly when built.
-      "@alextheman/no-relative-imports": "off",
-    },
-  },
-];
