+++ packages/prisma-client/src/prisma.ts
@@ -1,1 +1,1 @@
-export { PrismaClient } from "./generated/prisma";
+export * from "./generated/prisma";
