+++ packages/types/eslint.config.js
@@ -1,0 +1,3 @@
+import plugin from "@alextheman/eslint-plugin"
+
+export default plugin.configs.alexTypeScriptBase
